update migrate DB

+++ database/migrations/2020_08_02_122905_create_type_users_table.php
@@ -14,7 +14,7 @@ class CreateTypeUsersTable extends Migration
     public function up()
     {
         Schema::create('type_users', function (Blueprint $table) {
-            $table->unsignedInteger('user_id')->index();
+            $table->unsignedBigInteger('user_id')->index();
             $table->unsignedInteger('type_id');
         });
     }
